Refactor python live samples including partials
